- Improved the ui throughout the site:   - Added an ability to check the user invite     status without refreshing and adding the user     to the database   - Fixed many visual bugs - Improved verification of user input - Overall code refactor

+++ app.py
@@ -13,7 +13,6 @@ from models import StudySession, Participant, History
 from tortoise.contrib.quart import register_tortoise
 import pytz
 from datetime import datetime
-from datetime import timedelta
 
 datetime.utcnow().replace(tzinfo=pytz.utc)
 QUART_APP = Quart(__name__)
@@ -60,7 +59,8 @@ def after_request(response):
 async def homepage():
     if request.method == "POST":
         # Checks if the user is already a manager of a session
-        if not await StudySession.filter(manager=ses["user_id"]).exists():
+        if not await StudySession.filter(manager=ses["user_id"]).exists() \
+                or not await Participant.filter(discord_id=ses["user_id"]).exists():
             if "invite_list" in request.headers["X-Custom-Header"]:
                 payload = await request.get_json()
                 users_id = payload["users_id"]
@@ -69,8 +69,6 @@ async def homepage():
                     user_id = ses["user_id"]
                     user_info = ses["user_payload"]
                     user_info_username = user_info["username"]
-                    print(user_info_username)
-                    users_amount = len(users_id)
                     channel_name = payload["topic"]
                     desc = payload["desc"]
 
@@ -78,7 +76,7 @@ async def homepage():
                     dm_channel_id = [dm_channel["id"] for dm_channel in dm_channels]
                     voice_channel = await client.create_voice_channel(guild_id=guild_id,
                                                                       channel_name=channel_name,
-                                                                      users_limit=users_amount)
+                                                                      users_limit=len(users_id))
                     voice_channel_id = voice_channel["id"]
                     invite_msg = await client.create_invite(channel_id=voice_channel_id)
                     inv_msg_id = await client.inv_multiple_users(dm_channel_id, invite=invite_msg)
@@ -86,44 +84,39 @@ async def homepage():
                     # wrapping the session by a user id and his two values, msg_id and channel id
                     active_session_payload = dict(zip(users_id, list((zip(dm_channel_id, inv_msg_id)))))
                     ses["active_session_payload"] = active_session_payload
-                    ses["active_session_users"] = [ses["user_id"], ]
+                    ses["active_session_users"] = []
                     # stores in db
-                    await StudySession.create(name=channel_name if channel_name else "No Topic Study Room",
-                                              manager=user_id,
-                                              guild=guild_id,
-                                              manager_username=user_info_username,
-                                              voice_channel_id=voice_channel_id,
-                                              desc=desc)
+                    study_session = await StudySession.create(name=channel_name if channel_name
+                                                              else "No Topic Study Room",
+                                                              manager=user_id,
+                                                              guild=guild_id,
+                                                              manager_username=user_info_username,
+                                                              voice_channel_id=voice_channel_id,
+                                                              desc=desc if len(desc) > 0 else "No Description")
+                    await Participant.create(session_id=study_session.id, discord_id=ses["user_id"])
                 else:
                     return "Wrong user number"
 
-        if "start_session" in request.headers["X-Custom-Header"]:
-            query = await StudySession.filter(manager=ses["user_id"]).first()
-            # adds all ready users
-            for user in ses["active_session_users"]:
-                await Participant.create(session_id=query.id, discord_id=user)
-
         if "stop_session" in request.headers["X-Custom-Header"]:
-            query = await StudySession.filter(manager=ses["user_id"]).first()
-            participants = await Participant.filter(session=query.id).all()
-            current_time = datetime.now(pytz.utc)
+            if StudySession.filter(manager=ses["user_id"]):
+                query = await StudySession.filter(manager=ses["user_id"]).first()
+                participants = await Participant.filter(session=query.id).all()
+                current_time = datetime.now(pytz.utc)
+                for participant in participants:
+                    duration = current_time - participant.start
+                    # Change later to 300
+                    if duration.total_seconds() >= 5:
+                        await History.create(session_name=query.name,
+                                             desc=query.desc,
+                                             user_id=participant.discord_id,
+                                             duration=duration)
+                        # user_time = timedelta(seconds=history.duration.total_seconds())
 
-            for participant in participants:
-                duration = current_time - participant.start
-                # Change later to 300
-                if duration.total_seconds() >= 5:
-                    history = await History.create(session_name=query.name,
-                                                   desc=query.desc if query.desc is None else "No Description",
-                                                   user_id=participant.discord_id,
-                                                   duration=duration)
-                    user_time = timedelta(seconds=history.duration.total_seconds())
-                    print(user_time)
-
-            await client.delete_channel(channel_id=query.voice_channel_id)
-            await query.delete()
-            ses["active_session_payload"] = None
-            # need to remove the invite link too and maybe the message aswell
-            return redirect("/")
+                await client.delete_channel(channel_id=query.voice_channel_id)
+                await query.delete()
+                ses["active_session_payload"] = None
+                # need to remove the invite link too and maybe the message aswell
+                return redirect("/")
 
         if "leave_session" in request.headers["X-Custom-Header"]:
             user_id = ses["user_id"]
@@ -132,76 +125,99 @@ async def homepage():
             current_time = datetime.now(pytz.utc)
             duration = current_time - user_session.start
             await History.create(session_name=session_info.name,
-                                 desc=session_info.desc if session_info.desc is None else "No Description",
+                                 desc=session_info.desc if session_info.desc is not None else "No Description",
                                  user_id=user_session.discord_id,
                                  duration=duration)
             await user_session.delete()
             return redirect("/")
-        if "check_user" in request.headers["X-Custom-Header"]:
-            check_user = await request.get_json()
-            check_user_id = check_user["user_id"]
-            active_sessions = ses["active_session_payload"]
-            active_dm_session = active_sessions[check_user_id]
 
-            active_user = await client.check_reaction(active_dm_session)
-            if active_user[0]["id"] in active_sessions:
+        if "check_user" in request.headers["X-Custom-Header"]:
+            if StudySession.filter(manager=ses["user_id"]).exists:
+                # later store the active session somewhere else, maybe a cookie
+                query = await StudySession.filter(manager=ses["user_id"]).first()
+                check_user = await request.get_json()
+                check_user_id = check_user["user_id"]
+                active_sessions = ses["active_session_payload"]
+                active_dm_session = active_sessions[check_user_id]
                 active_session_users = ses["active_session_users"]
-                if active_user[0]["id"] not in active_session_users:
+
+                active_user = await client.check_reaction(active_dm_session)
+                if active_user[0]["id"] in active_sessions not in active_session_users:
+                    test = await Participant.create(session_id=query.id, discord_id=active_user[0]["id"])
                     active_session_users.append(active_user[0]["id"])
                     ses["active_session_users"] = active_session_users
-                return "Joined"
+                    return "Joined"
+                else:
+                    return "Invited"
             else:
-                return "Invited"
+                "You are not a session manager"
         return "hello"
 
     else:
         guild_id = ses["user_guild_id"]
         user_payload = ses["user_payload"]
         if ses.get("guild_users") is None:
+            # saving all the users in the chosen guild
             response = await client.get_guild_members(guild_id)
             guild_users = {
                 user["user"]["id"]: {
                     "username": user["user"]["username"],
                     "avatar": user["user"]["avatar"]}
-                for user in response if user["user"]["id"] != str(ses["user_id"])
-                and "bot" not in user["user"]}
+                for user in response if user["user"]["id"] != str(ses["user_id"]) and "bot" not in user["user"]}
             ses["guild_users"] = guild_users
 
         # manager view: need to work on var naming for manager and participant
         if await StudySession.filter(manager=ses["user_id"]).exists():
+            # need to reorder
+            session_manager = True
             session_info = await StudySession.filter(manager=ses["user_id"]).first()
             user_session = await Participant.filter(discord_id=ses["user_id"]).first()
             guild_users = ses["guild_users"]
-            active_session_users = await Participant.filter(session_id=user_session.session_id).all()
-            active_users_id = [str(active_user.discord_id) for active_user in active_session_users
-                               if active_user.discord_id != ses["user_id"]]
+            active_session_users = await Participant.exclude(discord_id=ses["user_id"]).filter(session_id=user_session.session_id).all()
+            # a list of the users id that are participating in the session without the user it self
+            active_users_id = [str(active_user.discord_id) for active_user in active_session_users]
             active_users_payload = {user: guild_users[user] for user in active_users_id if user in guild_users}
-            session_manager = True
+            active_ses = ses["active_session_payload"]
+            users_id = [active for active in active_ses]
+            users_info = {user: guild_users[user] for user in users_id}
             user_payload = ses["user_payload"]
             duration = datetime.now(pytz.utc) - user_session.start
             duration = str(duration).split(".")[0]
+            # check the vars again, and maybe better name them
+            #print(f"this is session_partic {user_session}\n this is session_partic by name {users_info}\n"
+                  #f" this is active_user partic {active_users_payload}")
             return await render_template("index.html", session_info=session_info,
                                          session_partic=user_session,
+                                         session_partic_byname=users_info,
                                          active_users_partic=active_users_payload,
                                          session_manager=session_manager,
                                          duration=duration,
+                                         manager=True,
                                          user_payload=user_payload)
 
         # participant view
         elif await Participant.filter(discord_id=ses["user_id"]).exists():
-            print("participant")
             user_session = await Participant.filter(discord_id=ses["user_id"]).first()
             session_info = await StudySession.filter(id=user_session.session_id).first()
+            # make sure that the stored guild users in the cookie is matching to the guild in the ses
+            # he's joining
+            if session_info.guild != ses["user_guild_id"]:
+                response = await client.get_guild_members(session_info.guild)
+                # refactor into a function
+                guild_users = {
+                    user["user"]["id"]: {
+                        "username": user["user"]["username"],
+                        "avatar": user["user"]["avatar"]}
+                    for user in response if user["user"]["id"] != str(ses["user_id"])
+                                            and "bot" not in user["user"]}
+                ses["guild_users"] = guild_users
             user_payload = ses["user_payload"]
             guild_users = ses["guild_users"]
-
-            active_session_users = await Participant.filter(session_id=user_session.session_id).all()
-            active_users_id = [str(active_user.discord_id) for active_user in active_session_users
-                               if active_user.discord_id != ses["user_id"]]
+            session_users = await Participant.exclude(discord_id=ses["user_id"]).filter(session_id=user_session.session_id).all()
+            active_users_id = [str(active_user.discord_id) for active_user in session_users]
             active_users_payload = {user: guild_users[user] for user in active_users_id if user in guild_users}
             duration = datetime.now(pytz.utc) - user_session.start
             formatted_dur = str(duration).split(".")[0]
-            print(user_payload)
             return await render_template("index.html",
                                          session_partic=user_session,
                                          duration=formatted_dur,
@@ -209,7 +225,9 @@ async def homepage():
                                          participant=True,
                                          user_payload=user_payload,
                                          session_info=session_info)
-        return await render_template("index.html", guild_users=ses["guild_users"], user_payload=user_payload)
+        return await render_template("index.html",
+                                     guild_users=ses["guild_users"],
+                                     user_payload=user_payload)
 
 
 @QUART_APP.route("/login/")
@@ -227,7 +245,6 @@ async def history():
     user_history = {query.index(study_session): {"name": study_session.session_name,
                                                  "duration": str(study_session.duration).split(".")[0],
                                                  "desc": study_session.desc} for study_session in query}
-    print(user_history)
     return await render_template("history.html", user_payload=ses["user_payload"],
                                  user_history=user_history)
 
@@ -237,7 +254,6 @@ async def guild():
     if request.method == "POST":
         if "guild_name" in await request.get_json(force=True):
             guild_name = await request.get_json(force=True)
-            print(guild_name)
             # I already got a list of guilds and stores it in session, better to use it again.
             # I can get the id from list of dicts much eaiser than what I did here. # (check example down "test")
             bot_guilds = await discord.bot_request("/users/@me/guilds", method="GET")
